feat: enhance services and controllers with improved documentation and WebSocket handling
