+++ lib/src/core/services/connectivity_service.dart
@@ -1,27 +1,36 @@
+/// Service for connectivity.
+/// Use streamSubscription to provide update connectivity status
+
+library;
+
 import 'dart:async';
 
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:get/get.dart';
 
+/// Service class for connectivity service.
+/// Check connectivity status in user device.
 class ConnectivityService extends GetxService {
   final RxBool isConnected = true.obs;
   late StreamSubscription<List<ConnectivityResult>> _subscription;
 
   @override
   void onInit() {
-    // TODO: implement onInit
     super.onInit();
     checkInitialConnection();
+    // create stream subscription that listen on connectivityChanged to updateConnectionStatus.
     _subscription = Connectivity().onConnectivityChanged.listen(
       _updateConnectionStatus,
     );
   }
 
+  /// Check initial device connection then send the result to [_updateConnectionStatus] function.
   Future<void> checkInitialConnection() async {
     final result = await Connectivity().checkConnectivity();
     _updateConnectionStatus(result);
   }
 
+  /// update connection status if conectivity result contains mobile, wifi, or ethernet.
   void _updateConnectionStatus(List<ConnectivityResult> result) {
     if (result.contains(ConnectivityResult.mobile) ||
         result.contains(ConnectivityResult.wifi) ||
